+++ src/main/java/techcourse/myblog/domain/Comment.java
@@ -7,6 +7,7 @@ import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.OnDelete;
 import org.hibernate.annotations.OnDeleteAction;
 import org.hibernate.annotations.UpdateTimestamp;
+import techcourse.myblog.support.exception.InvalidCommentException;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
