+++ src/main/java/techcourse/myblog/support/exception/IllegalUserException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.support.exception;
 
 public class IllegalUserException extends RuntimeException {
     public IllegalUserException(String msg) {
