+++ src/main/java/techcourse/myblog/support/exception/InvalidCommentException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.support.exception;
 
 public class InvalidCommentException extends RuntimeException {
     public InvalidCommentException(String msg) {
