+++ src/main/java/techcourse/myblog/support/exception/LoginFailedException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service;
+package techcourse.myblog.support.exception;
 
 public class LoginFailedException extends RuntimeException {
     public LoginFailedException(String message) {
