+++ src/main/java/techcourse/myblog/support/exception/MismatchAuthorException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service;
+package techcourse.myblog.support.exception;
 
 public class MismatchAuthorException extends RuntimeException {
     public MismatchAuthorException(String msg) {
