+++ src/main/java/techcourse/myblog/support/exception/NotFoundArticleException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service;
+package techcourse.myblog.support.exception;
 
 public class NotFoundArticleException extends RuntimeException {
     public NotFoundArticleException(String msg) {
