+++ src/main/java/techcourse/myblog/web/controller/UserControllerExceptionHandler.java
@@ -4,10 +4,10 @@ import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.springframework.web.servlet.view.RedirectView;
-import techcourse.myblog.domain.InvalidCommentException;
-import techcourse.myblog.service.DuplicatedEmailException;
-import techcourse.myblog.service.MismatchAuthorException;
-import techcourse.myblog.service.NotFoundArticleException;
+import techcourse.myblog.support.exception.InvalidCommentException;
+import techcourse.myblog.support.exception.DuplicatedEmailException;
+import techcourse.myblog.support.exception.MismatchAuthorException;
+import techcourse.myblog.support.exception.NotFoundArticleException;
 
 @ControllerAdvice
 public class UserControllerExceptionHandler {
