+++ src/test/java/techcourse/myblog/domain/UserTest.java
@@ -1,10 +1,10 @@
 package techcourse.myblog.domain;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import techcourse.myblog.support.exception.IllegalUserException;
 
 import java.util.stream.Stream;
