+++ src/test/java/techcourse/myblog/service/UserWriteServiceTests.java
@@ -3,6 +3,7 @@ package techcourse.myblog.service;
 import org.junit.jupiter.api.Test;
 import techcourse.myblog.service.dto.UserDto;
 import techcourse.myblog.service.common.UserCommonServiceTests;
+import techcourse.myblog.support.exception.DuplicatedEmailException;
 
 import java.util.Optional;
